ya se pueden agregar promos

+++ Neatburguer/Areas/Admin/Models/ViewModels/MenuAdminViewModel.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string Nombre { get; set; } = null!;
         public string Descripcion { get; set; } = null!;
-        public double Precio { get; set; }
-        public double PrecioPromocion { get; set; }
+        public decimal Precio { get; set; }
+        public decimal PrecioPromocion { get; set; }
 
         public string Clasificacion { get; set; } = null!;
